+++ MapleServer2/Triggers/FieldObjectTriggers.cs
@@ -170,8 +170,7 @@ public partial class TriggerContext
                 SkillCast skillCast = new(triggerSkill.Value.SkillId, triggerSkill.Value.SkillLevel, GuidGenerator.Long(), Environment.TickCount)
                 {
                     SkillObjectId = triggerSkill.ObjectId,
-                    //CasterObjectId = triggerSkill.ObjectId,
-                    Position = triggerSkill.Coord,
+                    Position = triggerSkill.Coord
                 };
                 RegionSkillHandler.HandleEffect(Field, skillCast);
             }
